Add CSS classes to fields to indicate placeholder value
Added default CSS class for placeholder values. Can be used to style the look of placeholders, typically displayed in a gray color tone. Note that this commit does not use classes to indicate the "changed" status, like the original stylefix branch. This is intentional as the "changed" indicates normal (non-placeholder) values and thus developers should not be forced to style all normal fields, but only placeholder values.